requirements.txt
Flask~=3.0.3
pymongo~=4.7.2
requests~=2.31.0
python-dotenv~=1.0.1
APScheduler~=3.10.4
pytz~=2024.1
iqoptionapi~=7.1.1
ntplib~=0.4.0